+++ R/arguments.R
@@ -204,3 +204,66 @@ make_xy_call <- function(object, target) {
 
   fit_call
 }
+
+## -----------------------------------------------------------------------------
+#' Execution-time data dimension checks
+#'
+#' For some tuning parameters, the range of values depend on the data
+#' dimensions (e.g. `mtry`). Some packages will fail if the parameter values are
+#' outside of these ranges. Since the model might receive resampled versions of
+#' the data, these ranges can't be set prior to the point where the model is
+#' fit.  These functions check the possible range of the data and adjust them
+#' if needed (with a warning).
+#'
+#' @param num_cols,num_rows The parameter value requested by the user.
+#' @param source A data frame for the data to be used in the fit. If the source
+#' is named "data", it is assumed that one column of the data corresponds to
+#' an outcome (and is subtracted off).
+#' @param offset A number subtracted off of the number of rows available in the
+#' data.
+#' @return An integer (and perhaps a warning).
+#' @examples
+
+#' nearest_neighbor(neighbors= 100) %>%
+#'   set_engine("kknn") %>%
+#'   set_mode("regression") %>%
+#'   translate()
+#'
+#' library(ranger)
+#' rand_forest(mtry = 2, min_n = 100, trees = 3) %>%
+#'   set_engine("ranger") %>%
+#'   set_mode("regression") %>%
+#'   fit(mpg ~ ., data = mtcars)
+#' @export
+min_cols <- function(num_cols, source) {
+  cl <- match.call()
+  src_name <- rlang::expr_text(cl$source)
+  if (cl$source == "data") {
+    p <- ncol(source) - 1
+  } else {
+    p <- ncol(source)
+  }
+  if (num_cols > p) {
+    msg <- paste0(num_cols, " columns were requested but there were ", p,
+                 " predictors in the data. ", p, " will be used.")
+    rlang::warn(msg)
+    num_cols <- p
+  }
+
+  as.integer(num_cols)
+}
+
+#' @export
+#' @rdname min_cols
+min_rows <- function(num_rows, source, offset = 0) {
+  n <- nrow(source)
+
+  if (num_rows > n - offset) {
+    msg <- paste0(num_rows, " samples were requested but there were ", n,
+                  " rows in the data. ", n - offset, " will be used.")
+    rlang::warn(msg)
+    num_rows <- n - offset
+  }
+
+  as.integer(num_rows)
+}